Add files via upload
In this project you can convert Foot/Feet to Meter or Inch, and convert Meter to Foot/Feet or Inch, and can also convert Inch to Foot/Feet or Meter. First of all what you do in this project is to input the value. Then select a conversion option which will be mentioned in the program while running it.  After that select an option to convert to and the area value will be printed out according to your requirement.